24166 Fixed review comment related to function test case AB#24166 AB#43930
