+++ UKHO.ExternalNotificationService.API/UKHO.ExternalNotificationService.API.FunctionalTests/Helper/FssEventDataBase.cs
@@ -1,0 +1,59 @@
+﻿using Newtonsoft.Json.Linq;
+using System;
+using System.Collections.Generic;
+using UKHO.ExternalNotificationService.API.FunctionalTests.Model;
+
+namespace UKHO.ExternalNotificationService.API.FunctionalTests.Helper
+{
+    public static class FssEventDataBase
+    {
+        public static JObject GetFssEventBodyData()
+        {
+            var ensWebhookJson = JObject.Parse(@"{""Type"":""uk.gov.UKHO.FileShareService.NewFilesPublished.v1""}");
+            ensWebhookJson["Source"] = "https://files.admiralty.co.uk";
+            ensWebhookJson["Id"] = "49c67cca-9cca-4655-a38e-583693af55ea";
+            ensWebhookJson["Subject"] = "83d08093-7a67-4b3a-b431-92ba42feaea0";
+            ensWebhookJson["DataContentType"] = "application/json";
+            ensWebhookJson["Data"] = JObject.FromObject(GetFssEventData());
+
+            return ensWebhookJson;
+        }
+        public static FssEventData GetFssEventData()
+        {
+            Link linkBatchDetails = new()
+            {
+                Href = @"https://files.admiralty.co.uk/batch/83d08093-7a67-4b3a-b431-92ba42feaea0"
+            };
+            Link linkBatchStatus = new()
+            {
+                Href = @"https://files.admiralty.co.uk/batch/83d08093-7a67-4b3a-b431-92ba42feaea0/status"
+            };
+
+            FileLinks fileLinks = new()
+            {
+                Get = new Link() { Href = @"https://files.admiralty.co.uk/batch/83d08093-7a67-4b3a-b431-92ba42feaea0/files/AVCS_S631-1_Update_Wk45_21_Only.zip" },
+            };
+
+            BatchLinks links = new()
+            {
+                BatchDetails = linkBatchDetails,
+                BatchStatus = linkBatchStatus
+            };
+
+            return new FssEventData()
+            {
+                Links = links,
+                BusinessUnit = "AVCSData",
+                Attributes = new List<Attribute> { },
+                BatchId = "83d08093-7a67-4b3a-b431-92ba42feaea0",
+                BatchPublishedDate = DateTime.UtcNow,
+                Files = new BatchFile[] {new() { MIMEType= "application/zip",
+                                                                    FileName= "AVCS_S631-1_Update_Wk45_21_Only.zip",
+                                                                    FileSize=99073923,
+                                                                    Hash="yNpJTWFKhD3iasV8B/ePKw==",
+                                                                    Attributes=new List<Attribute> {},
+                                                                    Links = fileLinks   }}
+            };
+        }
+    }
+}
